% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_api_url.R
\name{create_api_url}
\alias{create_api_url}
\title{Create API URL}
\usage{
create_api_url(
  symbol,
  api_fun,
  api_key = "4HQVKYYYNZTDVUJ2",
  data_type = "json",
  ...
)
}
\arguments{
\item{symbol}{character object, ticker name}

\item{api_fun}{character object, api function defined by alphavantage}

\item{api_key}{character object, api key from alphavantage}

\item{data_type}{character object, either json or csv}
}
\value{
a url contains API request information
}
\description{
Create API URL to send request to alphavantage
}
