% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_target_stocks.R
\name{find_target_stocks}
\alias{find_target_stocks}
\title{Find Target Stock}
\usage{
find_target_stocks(symbol_list, strategy = "f-score")
}
\arguments{
\item{symbol_list}{vector of characters, a list of tickers}

\item{strategy}{character object, strategy to be used}
}
\value{
a list of tickers satisfy the strategy
}
\description{
Find target stock by different strategy
}
