% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_f_score_statistics.R
\name{get_f_score_statistics}
\alias{get_f_score_statistics}
\title{Get F-Score Statistics}
\usage{
get_f_score_statistics(symbol)
}
\arguments{
\item{symbol}{character object, ticker name}

\item{api_key}{character object, api key from alphavantage}

\item{data_type}{character object, either json or csv}
}
\value{
a list object contains f-score statistics
}
\description{
Get F-Score Statistics based on calculation
}
