% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_api_content.R
\name{get_raw_api_content}
\alias{get_raw_api_content}
\title{Get Raw API Content}
\usage{
get_raw_api_content(
  symbol,
  api_fun,
  api_key = "8ZQHIMUKAYE89GEH",
  data_type = "json",
  ...
)
}
\arguments{
\item{symbol}{character object, ticker name}

\item{api_fun}{character object, api function defined by alphavantage}

\item{api_key}{character object, api key from alphavantage}

\item{data_type}{character object, either json or csv}
}
\value{
raw content from alphavantage API
}
\description{
Get Raw API Content from alphavantage
}
